upgrade JSONHolder
