+++ app/src/main/java/com/example/laletin/picturerbog/DownloadPreviewTask.kt
@@ -6,6 +6,10 @@ import android.os.AsyncTask
 import android.util.Log
 
 
+interface onPreviewDownloaded {
+    fun onPreviewDownloaded()
+}
+
 class DownloadPreviewTask(private val holder: ImageViewHolder) : AsyncTask<Int, Void, Pair<Bitmap?, String?>?>() {
     override fun onPreExecute() {
         super.onPreExecute()
@@ -15,7 +19,7 @@ class DownloadPreviewTask(private val holder: ImageViewHolder) : AsyncTask<Int, 
     override fun doInBackground(vararg params: Int?): Pair<Bitmap?, String?>? {
         try {
             val index = params[0] ?: return null
-            val url_s = JSONHolder.json?.photos?.photo?.get(index)?.url_s
+            val url_s = JSONHolder().get()?.photos?.photo?.get(index)?.url_s
             var mIcon11: Bitmap? = null
 
             if (CacheForPreview.Cache.get(url_s) == null) {
@@ -34,7 +38,7 @@ class DownloadPreviewTask(private val holder: ImageViewHolder) : AsyncTask<Int, 
             } else {
                 mIcon11 = CacheForPreview.Cache.get(url_s)
             }
-            return Pair(mIcon11, JSONHolder.json?.photos?.photo?.get(index)?.title)
+            return Pair(mIcon11, JSONHolder().get()?.photos?.photo?.get(index)?.title)
         } catch (e: Exception) {
             e.printStackTrace()
         }
