+++ app/src/main/java/com/example/laletin/picturerbog/ImageList.kt
@@ -73,5 +73,4 @@ class ImageList : Fragment() {
             mBound = false
         }
     }
-
 }
